+++ STACpopulator/api_requests.py
@@ -1,6 +1,6 @@
 import logging
 import os
-from typing import Any, Optional
+from typing import Any, MutableMapping, Optional
 
 import requests
 from colorlog import ColoredFormatter
@@ -35,7 +35,7 @@ def stac_collection_exists(stac_host: str, collection_id: str) -> bool:
     return r.status_code == 200
 
 
-def post_stac_collection(stac_host: str, json_data: dict[str, Any], update: Optional[bool] = True) -> None:
+def post_stac_collection(stac_host: str, json_data: MutableMapping[str, Any], update: Optional[bool] = True) -> None:
     """Post/create a collection on the STAC host
 
     :param stac_host: address of the STAC host
@@ -62,7 +62,11 @@ def post_stac_collection(stac_host: str, json_data: dict[str, Any], update: Opti
 
 
 def post_stac_item(
-    stac_host: str, collection_id: str, item_name: str, json_data: dict[str, dict], update: Optional[bool] = True
+    stac_host: str,
+    collection_id: str,
+    item_name: str,
+    json_data: MutableMapping[str, dict],
+    update: Optional[bool] = True,
 ) -> None:
     """Post a STAC item to the host server.