+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-## [Unreleased](https://github.com/crim-ca/stac-populator) (latest)"
+## [Unreleased](https://github.com/crim-ca/stac-populator) (latest)
 
 * Add `LICENSE` file.
 * Add `bump-my-version` with `make version` and `make VERSION=<...> bump` utilities to self-update release versions.
@@ -10,6 +10,8 @@
 * Add `dev` dependencies to `pyproject.toml` for testing the package (install with `pip install ".[dev]"`).
 * Add GitHub CI tests.
 * Remove `requirements.txt` in favor of all dependencies combined in `pyproject.toml`.
+* Add test to validate STAC Collection and Item contain `source` with expected THREDDS format.
+* Fix broken tests and invalid imports.
 
 ## [0.1.0](https://github.com/crim-ca/stac-populator/tree/0.1.0) (2023-11-08)